Level Gen v2
Not sure where I am supposed to be pushing this but when I sent it too the master and then pulled it to test again, everything broke. So here is a messier throw together of part of what I had. Will put together the generation on each row in class.